Add remote client port in ConnectionAlert window. Add own class for connection alert data.
