+++ FSAClient/Classes/ConnectionAlertData.cs
@@ -1,0 +1,12 @@
+﻿namespace FSAClient.Classes
+{
+    public class ConnectionAlertData
+    {
+        public string UserName { get; set; }
+        public int UserId { get; set; }
+        public string IpAddress { get; set; }
+        public int Port { get; set; }
+        public string FileName { get; set; }
+        public string FileSize { get; set; }
+    }
+}
